Blog posts about rewrite, rework examples page
I wrote a blog post about the rewritten workshop in the main blog, and a
blog post about the resorted examples in the Grid Framework blog. While
I was at it, I also cleaned up the page by moving some of the content to
the individual example pages. I cleared up the CSS a bit and moved the
assets of the examples up a directory, next to the example page.