Fix paginator in the blog
The first issue was that a procedure (`page`) was being passed instead of an
integer (`current-page`). The second issue was a missing backslash in
the URL of a pager item, this cause the relative URL to go one directory
too far up (probably only affects the local server, but still).